Ver 0.1.2.8/0.0.0.7 Changed LongPollControl (ZHA) setting control algorithm and measurement interval (limit 3..30 seconds). For MHO-C401N added E-Ink refresh after 1024 display updates.

+++ src/lcd.h
@@ -52,7 +52,7 @@ typedef struct  {
 	u8 init;
 	u8 stage;
 	u8 updated;
-	u8 refresh_cnt;
+	u16 refresh_cnt;
 	u8 display_buff[LCD_BUF_SIZE];
 	u8 display_cmp_buff[LCD_BUF_SIZE];
 } scr_data_t;
